Add snake server side
