+++ Assets/Scripts/JustPlanes/Core/TestPlane.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0925504b51f6dd344912855d6981197a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
