+++ Assets/Scripts/JustPlanes/Unity/Snake.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a56a8f625c4697147a1e9a2ebe973b24
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
